requirements.txt
streamlit
torch
torchvision
transformers
pillow
pandas
tqdm
matplotlib
plotly
requests
